fix error in mada payment where the payment was taking long not double add method for devices to have the line string value remove spacing to be decreased from the devices changing the namings in order receipt and z report fix errors in calculations in the z report remove redundant commented code fix error in using the custom item product if description is not set put the username instead of user mail in z report and end shift summary

+++ app/src/main/java/com/app/smartpos/pos/ProductCart.java
@@ -671,7 +671,7 @@ public class ProductCart extends BaseActivity {
 
                 if(dialogOrderPaymentMethod.equals("CARD")) {
                     databaseAccess.open();
-                    long totalPriceWithTax = (long) databaseAccess.getTotalPriceWithTax();
+                    double totalPriceWithTax = databaseAccess.getTotalPriceWithTax();
 
                     Intent intent=device.pay(totalPriceWithTax);
 //                    if(){
